Move unused bin files to cli, call as function
